First stab at smarter report generation
The new rich reporting feature allows customizing reports (and failures) according to the result of the analysis.
It is possible to define a custom renderer, but the default one will use an embedded groovy template engine. Results
can be tweaked thanks to rules, which apply to changes (or non changes). For example, it is possible to express
complex rules like "this method has been added, but wasn't annotated with `@Incubating`").