Added default rules
The rich report will now never be empty if no rules are configured. It will instead apply "reasonable" rules which
will render binary breaking changes as errors, and report source compatible changes either as info (for additions)
or warnings (for other cases).

Execution order of rules is now changed, and it is possible to record if a rule has triggered a violation for a specific
member, so that it's not added multiple times for the same member (this could happen if you had a change rule and a generic
rule applied on the same member)